requirements.txt
# Core dependencies
pandas==2.1.3
neo4j==5.14.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
tqdm==4.66.1
numpy==1.24.3

# NLP/ML libraries
transformers==4.35.2
torch==2.1.0
sentencepiece==0.1.99
spacy==3.7.2

# Neo4j integration
neo4j-driver==5.14.0

# Optional: for web interface
streamlit==1.28.1

# Development
jupyter==1.0.0
ipython==8.17.2